move component/configuration API responses into client.py
