+++ integtests/test_validate.py
@@ -15,8 +15,7 @@ from typing import cast
 import jsonschema
 import pytest
 
-from keboola_mcp_server.client import JsonDict, KeboolaClient
-from keboola_mcp_server.tools.components.api_models import ComponentAPIResponse
+from keboola_mcp_server.client import ComponentAPIResponse, JsonDict, KeboolaClient
 from keboola_mcp_server.tools.components.model import Component
 from keboola_mcp_server.tools.validation import KeboolaParametersValidator
